+++ src/Hooks/UIPOptionsLayer.hpp
@@ -10,11 +10,20 @@ class $modify(ModUIPOptionsLayer, UIPOptionsLayer) {
 		SwitcherMenu* m_switcherMenu = nullptr;
 		bool m_movingSwitcher;
 		CCPoint m_lastPos;
+
+		TextInput* m_switcherScaleInput;
+		Slider* m_switcherScaleSlider;
 	};
 
 	bool init();
 	void onClose(CCObject* sender);
 	void onReset(CCObject* sender);
 
-	void saveSwitcherPosition();
+	void updateSwitcherScale(float scale);
+	void updateSwitcherScaleSlider(float scale);
+	void saveSwitcherSettings();
+
+	void onSwitcherScaleSliderUpdated(CCObject* slider);
 };
+
+float inverseLerp(float a, float b, float v);