+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.1.0
+- Added Android support
+- Added scale options for the switcher in the Practice Options menu
+- Fix practice buttons not following the opacity setting properly
+- Fixed a potential crash with mods that start levels in practice mode
+
 # 1.0.2
 Fix Practice UI Config dragging the layer instead of the switcher